+++ order_service/src/main/java/on/ssgdeal/order_service/application/service/SlackService.java
@@ -1,7 +1,7 @@
 package on.ssgdeal.order_service.application.service;
 
-import on.ssgdeal.order_service.infrastructure.client.slack.feign.dto.OrderCompleteSendSlackRequestDto;
-import on.ssgdeal.order_service.infrastructure.client.slack.feign.dto.OrderCompleteSendSlackResponseDto;
+import on.ssgdeal.order_service.infrastructure.client.slack.feign.dtos.OrderCompleteSendSlackRequestDto;
+import on.ssgdeal.order_service.infrastructure.client.slack.feign.dtos.OrderCompleteSendSlackResponseDto;
 
 public interface SlackService {
