+++ order_service/src/main/java/on/ssgdeal/order_service/application/service/UserService.java
@@ -1,7 +1,7 @@
 package on.ssgdeal.order_service.application.service;
 
-import on.ssgdeal.order_service.infrastructure.client.user.feign.dto.ValidDestinationRequestDto;
-import on.ssgdeal.order_service.infrastructure.client.user.feign.dto.ValidDestinationResponseDto;
+import on.ssgdeal.order_service.infrastructure.client.user.feign.dtos.ValidDestinationRequestDto;
+import on.ssgdeal.order_service.infrastructure.client.user.feign.dtos.ValidDestinationResponseDto;
 
 public interface UserService {
