+++ order_service/src/main/java/on/ssgdeal/order_service/infrastructure/client/promotion/feign/dtos/DecreaseProductStockRequestDto.java
@@ -1,4 +1,4 @@
-package on.ssgdeal.order_service.infrastructure.client.promotion.feign.dto;
+package on.ssgdeal.order_service.infrastructure.client.promotion.feign.dtos;
 
 public record DecreaseProductStockRequestDto(
     Long productId,
