+++ order_service/src/main/java/on/ssgdeal/order_service/infrastructure/client/slack/feign/SlackServiceFeignClient.java
@@ -1,8 +1,8 @@
 package on.ssgdeal.order_service.infrastructure.client.slack.feign;
 
 import on.ssgdeal.common.presentation.dto.CommonResponse;
-import on.ssgdeal.order_service.infrastructure.client.slack.feign.dto.OrderCompleteSendSlackRequestDto;
-import on.ssgdeal.order_service.infrastructure.client.slack.feign.dto.OrderCompleteSendSlackResponseDto;
+import on.ssgdeal.order_service.infrastructure.client.slack.feign.dtos.OrderCompleteSendSlackRequestDto;
+import on.ssgdeal.order_service.infrastructure.client.slack.feign.dtos.OrderCompleteSendSlackResponseDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
