+++ order_service/src/main/java/on/ssgdeal/order_service/infrastructure/client/slack/feign/dtos/OrderCompleteSendSlackRequestDto.java
@@ -1,10 +1,10 @@
-package on.ssgdeal.order_service.infrastructure.client.slack.feign.dto;
+package on.ssgdeal.order_service.infrastructure.client.slack.feign.dtos;
 
 import java.time.LocalDate;
 import on.ssgdeal.order_service.application.service.dto.LoginUserInfoDto;
 import on.ssgdeal.order_service.domain.entity.dtos.UpdateTotalOrderSuccessDto;
 import on.ssgdeal.order_service.domain.enums.TotalOrderStatus;
-import on.ssgdeal.order_service.infrastructure.client.slack.dto.TotalOrderCompleteSendInfoDto;
+import on.ssgdeal.order_service.infrastructure.client.slack.dtos.TotalOrderCompleteSendInfoDto;
 
 public record OrderCompleteSendSlackRequestDto(Long totalOrderId, String slackEmail,
                                                LocalDate OrderAt, Long paymentPrice,
