+++ order_service/src/main/java/on/ssgdeal/order_service/infrastructure/client/slack/feign/dtos/OrderCompleteSendSlackResponseDto.java
@@ -1,4 +1,4 @@
-package on.ssgdeal.order_service.infrastructure.client.slack.feign.dto;
+package on.ssgdeal.order_service.infrastructure.client.slack.feign.dtos;
 
 public record OrderCompleteSendSlackResponseDto(Long notificationId) {
