+++ order_service/src/main/java/on/ssgdeal/order_service/infrastructure/client/user/feign/UserServiceFeignClient.java
@@ -1,8 +1,8 @@
 package on.ssgdeal.order_service.infrastructure.client.user.feign;
 
 import on.ssgdeal.common.presentation.dto.CommonResponse;
-import on.ssgdeal.order_service.infrastructure.client.user.feign.dto.ValidDestinationRequestDto;
-import on.ssgdeal.order_service.infrastructure.client.user.feign.dto.ValidDestinationResponseDto;
+import on.ssgdeal.order_service.infrastructure.client.user.feign.dtos.ValidDestinationRequestDto;
+import on.ssgdeal.order_service.infrastructure.client.user.feign.dtos.ValidDestinationResponseDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
